提交

+++ apartment-sartup/src/main/java/com/link/model/User.java
@@ -3,7 +3,9 @@ package com.link.model;
 import lombok.Getter;
 import lombok.Setter;
 
+
 import javax.persistence.*;
+
 
 @Entity
 @Getter
@@ -12,11 +14,14 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-        private Long id;
+    private int id;
     @Column(name = "name", nullable = true, length = 30)
-        private String name;
+    private String name;
     @Column(name = "password", nullable = true, length = 30)
-        private String password;
-    }
+    private String password;
+
+
+
+}
